Members of German parliament that represent districts with universities
This repository provides you with the tools to find out which members of the German parliament represent districts with at least one university. The lists also contain their email-addresses.
I included the final csv files for your convenience in the output folder.
You will need to have a Google Maps API if you want to run the code from scratch. However I provide you with an intermediate file in the output folder. This way you can skip the geocoding step.